refactor: clippy fixes

+++ crates/ironboyadvance_arm7tdmi/src/arm/execute.rs
@@ -28,7 +28,7 @@ pub fn execute_branch_and_branch_link<I: MemoryInterface>(
         cpu.set_register(LR, cpu.pc() - 4)
     }
 
-    let offset = (((instruction.offset() as u32) << 8) as i32) >> 6;
+    let offset = ((instruction.offset() << 8) as i32) >> 6;
     cpu.set_pc((cpu.pc() as i32).wrapping_add(offset) as u32);
     cpu.pipeline_flush();
     CpuAction::PipelineFlush
